+++ src/shared/prisma/migrations/20260426032847_add_user_model/migration.sql
@@ -1,0 +1,23 @@
+-- AlterTable
+ALTER TABLE "Recipe" ALTER COLUMN "thumbnail" DROP NOT NULL,
+ALTER COLUMN "thumbnail" SET DEFAULT 'https://placehold.co/900x600/1C4532/C6F6D5.png?text=Imagen+de\nla+Receta';
+
+-- CreateTable
+CREATE TABLE "User" (
+    "id" SERIAL NOT NULL,
+    "email" TEXT NOT NULL,
+    "username" TEXT NOT NULL,
+    "passwordHash" TEXT NOT NULL,
+    "isActive" BOOLEAN NOT NULL DEFAULT true,
+    "lastLoginAt" TIMESTAMP(3),
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+
+    CONSTRAINT "User_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "User_email_key" ON "User"("email");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "User_username_key" ON "User"("username");