+++ src/shared/prisma/migrations/20260426043922_add_user_fk_to_recipe/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Recipe" ADD COLUMN     "idUser" INTEGER;
+
+-- AddForeignKey
+ALTER TABLE "Recipe" ADD CONSTRAINT "Recipe_idUser_fkey" FOREIGN KEY ("idUser") REFERENCES "User"("id") ON DELETE SET NULL ON UPDATE CASCADE;